+++ lib/integration/lib/model/create-integration-instance-details.ts
@@ -87,6 +87,14 @@ export interface CreateIntegrationInstanceDetails {
    * Shape
    */
   "shape"?: CreateIntegrationInstanceDetails.Shape;
+  /**
+   * The OCID of the identity domain, that will be used to determine the
+   * corresponding Idcs Stripe and create an Idcs application within the stripe.
+   * This parameter is mutually exclusive with parameter: idcsAt, i.e only one of
+   * two parameters should be specified.
+   *
+   */
+  "domainId"?: string;
 }
 
 export namespace CreateIntegrationInstanceDetails {