+++ lib/keymanagement/index.ts
@@ -1,7 +1,6 @@
 /**
- * Vault Service Key Management API
- * API for managing and performing operations with keys and vaults. (For the API for managing secrets, see the Vault Service 
-Secret Management API. For the API for retrieving secrets, see the Vault Service Secret Retrieval API.)
+ * Vault Key Management API
+ * Use the Key Management API to manage vaults and keys. For more information, see [Managing Vaults](/Content/KeyManagement/Tasks/managingvaults.htm) and [Managing Keys](/Content/KeyManagement/Tasks/managingkeys.htm).
 
  * OpenAPI spec version: release
  * Contact: sparta_kms_us_grp@oracle.com
@@ -17,12 +16,15 @@ import * as requests from "./lib/request";
 import * as models from "./lib/model";
 import * as responses from "./lib/response";
 import * as client from "./lib/client";
+import * as ekm_waiter from "./lib/ekm-waiter";
 import * as kmsmanagement_waiter from "./lib/kmsmanagement-waiter";
 import * as kmsvault_waiter from "./lib/kmsvault-waiter";
 
 export { models };
 export { requests };
 export { responses };
+export import EkmClient = client.EkmClient;
+export import EkmWaiter = ekm_waiter.EkmWaiter;
 export import KmsCryptoClient = client.KmsCryptoClient;
 export import KmsManagementClient = client.KmsManagementClient;
 export import KmsManagementWaiter = kmsmanagement_waiter.KmsManagementWaiter;